Use cert verifier `GetRecencyWindow`
* Changes existing recency window types to uint32 to align with contract
* Uses GetRecencyWindow to fetch value from contract, or fallback to flag value

+++ api/proxy/common/client_config_v2.go
@@ -49,7 +49,7 @@ type ClientConfigV2 struct {
 	// See https://layr-labs.github.io/eigenda/integration/spec/6-secure-integration.html#1-rbn-recency-validation
 	//
 	// This check is optional and will be skipped when RBNRecencyWindowSize is set to 0.
-	RBNRecencyWindowSize uint64
+	RBNRecencyWindowSize uint32
 
 	// The EigenDA network that is being used.
 	// It is optional, and when set will be used for validating that the eth-rpc chain ID matches the network.
